Use integer for protocol for custom protocol implementation

+++ JKClient/Handlers/JA/JANetHandler.cs
@@ -1,6 +1,6 @@
 ﻿namespace JKClient {
 	public abstract class JANetHandler : NetHandler {
 		public override int MaxMessageLength => 49152;
-		public JANetHandler(ProtocolVersion protocol) : base(protocol) {}
+		public JANetHandler(ProtocolVersion protocol) : base((int)protocol) {}
 	}
 }
